+++ lib/src/script.dart
@@ -54,6 +54,8 @@ class VMScriptRef implements VMObjectRef {
       new VMScript._(_scope, await _scope.loadObject(_id));
 
   /// Adds a breakpoint at [line] (and optionally [column]) in this script.
+  ///
+  /// Both [line] and [column] are 1-based.
   Future<VMBreakpoint> addBreakpoint(int line, {int column}) async {
     var params = {"scriptId": _id, "line": line};
     if (column != null) params["column"] = column;
@@ -108,6 +110,9 @@ class VMScript extends VMScriptRef implements VMObject {
   /// This is generally less efficient than calling [sourceSpan] and
   /// [sourceLocation] directly, and should only be used when you don't
   /// [VMSourceLocation] or [VMScriptToken] objects.
+  ///
+  /// Note that [SourceFile] uses 0-based lines and columns, whereas the rest of
+  /// this package uses 1-based lines and columns.
   SourceFile get sourceFile {
     if (_sourceFile == null) _sourceFile = new SourceFile(source, url: uri);
     return _sourceFile;
@@ -125,7 +130,8 @@ class VMScript extends VMScriptRef implements VMObject {
   /// Returns a [FileSpan] representing the source covered by [location].
   ///
   /// If [location] doesn't have a [VMSourceLocation.end] token, this will be a
-  /// point span.
+  /// point span. Note that [FileSpan] uses 0-based lines and columns, whereas
+  /// the rest of this package uses 1-based lines and columns.
   ///
   /// Throws an [ArgumentError] if [location] isn't for this script.
   FileSpan sourceSpan(VMSourceLocation location) {
@@ -135,10 +141,13 @@ class VMScript extends VMScriptRef implements VMObject {
     }
 
     var end = location.end ?? location.token;
-    return new _ScriptSpan(this, location.token.offset, end.offset);
+    return new _ScriptSpan(this, location.token._position, end._position);
   }
 
   /// Returns a [FileLocation] representing the location indicated by [token].
+  ///
+  /// Note that [FileLocation] uses 0-based lines and columns, whereas the rest
+  /// of this package uses 1-based lines and columns.
   ///
   /// Throws an [ArgumentError] if [location] isn't for this script.
   FileLocation sourceLocation(VMScriptToken token) {
@@ -147,15 +156,15 @@ class VMScript extends VMScriptRef implements VMObject {
       throw new ArgumentError("Token isn't for this script.");
     }
 
-    return new _ScriptLocation(this, token.offset);
+    return new _ScriptLocation(this, token._position);
   }
 
   /// Binary searches [_tokenPositions] for the line and column information for
-  /// the token at [offset].
+  /// the token at [position].
   ///
   /// This returns a `line, column` pair if the token is found, and throws a
   /// [StateError] otherwise.
-  List<int> _lineAndColumn(int offset) {
+  List<int> _lineAndColumn(int position) {
     var min = 0;
     var max = _tokenPositions.length;
     while (min < max) {
@@ -163,11 +172,11 @@ class VMScript extends VMScriptRef implements VMObject {
 
       var row = _tokenPositions[mid];
 
-      if (row[1] > offset) {
+      if (row[1] > position) {
         max = mid;
       } else {
         for (var i = 1; i < row.length; i += 2) {
-          if (row[i] == offset) return [row.first, row[i + 1]];
+          if (row[i] == position) return [row.first, row[i + 1]];
         }
 
         min = mid + 1;
@@ -176,7 +185,7 @@ class VMScript extends VMScriptRef implements VMObject {
 
     // We only call this for positions that come from the VM, so we shouldn't
     // ever actually reach this point.
-    throw new StateError("Couldn't find line and column for offset $offset in "
+    throw new StateError("Couldn't find line and column for token $position in "
         "$uri.");
   }
 }
@@ -192,21 +201,35 @@ class VMScriptToken {
   /// The ID of this token's script.
   final String _scriptId;
 
-  /// The location of this token in the script, in code units from the
-  /// beginning.
-  final int offset;
+  /// This is deprecated.
+  ///
+  /// Use `VMScript.sourceLocation().offset` instead.
+  @Deprecated("Will be removed in 0.3.0.")
+  int get offset => _position;
+  final int _position;
 
-  VMScriptToken._(this._isolateId, this._scriptId, this.offset);
+  VMScriptToken._(this._isolateId, this._scriptId, this._position);
 
-  String toString() => offset.toString();
+  String toString() => _position.toString();
 }
 
-/// An implementation of [FileLocation] based on a known token offset.
+/// An implementation of [FileLocation] based on a known token position.
 class _ScriptLocation extends SourceLocationMixin implements FileLocation {
   /// The script that produced this location.
   final VMScript _script;
 
-  final int offset;
+  /// The position of the token used to generate this location.
+  ///
+  /// This is opaque, but it can be used to look up the line and column
+  /// information.
+  final int _position;
+
+  int get offset {
+    // TODO(nweiz): Make this more efficient when sdk#26390 is fixed.
+    _offset ??= _script.sourceFile.getOffset(line, column);
+    return _offset;
+  }
+  int _offset;
 
   SourceFile get file => _script.sourceFile;
 
@@ -224,78 +247,106 @@ class _ScriptLocation extends SourceLocationMixin implements FileLocation {
   }
   int _column;
 
-  _ScriptLocation(this._script, this.offset);
+  _ScriptLocation(this._script, this._position);
 
   /// Ensures that [_line] and [_column] are set based on [_script]'s
   /// information.
   void _ensureLineAndColumn() {
     if (_line != null) return null;
-    var result = _script._lineAndColumn(offset);
-    _line = result.first;
-    _column = result.last;
+    var result = _script._lineAndColumn(_position);
+
+    // SourceLocation lines and columns are zero-based, the VM service's are
+    // 1-based.
+    _line = result.first - 1;
+    _column = result.last - 1;
   }
 
-  FileSpan pointSpan() => new _ScriptSpan(_script, offset, offset);
+  int compareTo(SourceLocation other) =>
+      other is _ScriptLocation && other.sourceUrl == sourceUrl
+          ? _position - other._position
+          : super.compareTo(other);
+
+  bool operator ==(SourceLocation other) => other is _ScriptLocation
+      ? _position == other._position && sourceUrl == other.sourceUrl
+      : super == other;
+
+  FileSpan pointSpan() =>
+      new _ScriptSpan(_script, _position, _position, this, this);
 }
 
-/// An implementation of [FileSpan] based on known token offsets.
+/// An implementation of [FileSpan] based on known token positions.
 class _ScriptSpan extends SourceSpanMixin implements FileSpan {
   /// The script that produced this location.
   final VMScript _script;
 
   SourceFile get file => _script.sourceFile;
 
-  /// The offset of the start token.
-  final int _start;
+  /// The position of the start token.
+  final int _startPosition;
 
-  /// The offset of the end token.
-  final int _end;
+  /// The position of the end token.
+  final int _endPosition;
 
   Uri get sourceUrl => _script.uri;
-  int get length => _end - _start;
-  FileLocation get start => new _ScriptLocation(_script, _start);
-  FileLocation get end => new _ScriptLocation(_script, _end);
-  String get text => _script.source.substring(_start, _end);
 
-  // We have to make [_script.sourceFile] concrete for this because the VMScript
-  // alone doesn't have the line context in an easily accessible form.
+  int get length {
+    // Prefer checking just the lines and columns, since they're more efficient
+    // to access.
+    if (start.line == end.line) return end.column - start.column;
+    return end.offset - start.offset;
+  }
+
+  FileLocation get start {
+    _start ??= new _ScriptLocation(_script, _startPosition);
+    return _start;
+  }
+  FileLocation _start;
+
+  FileLocation get end {
+    _end ??= new _ScriptLocation(_script, _endPosition);
+    return _end;
+  }
+  FileLocation _end;
+
+  String get text => _script.source.substring(start.offset, end.offset);
+
   String get context => file.getText(file.getOffset(start.line),
       end.line == file.lines - 1 ? null : file.getOffset(end.line + 1));  
 
-  _ScriptSpan(this._script, this._start, this._end);
+  _ScriptSpan(this._script, this._startPosition, this._endPosition,
+      [this._start, this._end]);
 
   int compareTo(SourceSpan other) {
-    if (other is! _ScriptSpan) return super.compareTo(other);
+    if (other is! _ScriptSpan || other.sourceUrl != sourceUrl) {
+      return super.compareTo(other);
+    }
 
     _ScriptSpan otherFile = other;
-    var result = _start.compareTo(otherFile._start);
-    return result == 0 ? _end.compareTo(otherFile._end) : result;
+    var result = _startPosition.compareTo(otherFile._startPosition);
+    return result == 0 ?
+        _endPosition.compareTo(otherFile._endPosition)
+        : result;
   }
 
   SourceSpan union(SourceSpan other) {
     if (other is! _ScriptSpan) return super.union(other);
 
     var span = expand(other);
-    var beginSpan = span._start == _start ? this : other;
-    var endSpan = span._end == _end ? this : other;
+    var beginSpan = span._startPosition == _startPosition ? this : other;
+    var endSpan = span._endPosition == _endPosition ? this : other;
 
-    if (beginSpan._end < endSpan._start) {
+    if (beginSpan._endPosition < endSpan._startPosition) {
       throw new ArgumentError("Spans $this and $other are disjoint.");
     }
 
     return span;
   }
 
-  bool operator ==(other) {
-    if (other is! FileSpan) return super == other;
-
-    if (other is! _ScriptSpan) {
-      return super == other && sourceUrl == other.sourceUrl;
-    }
-
-    return _start == other._start && _end == other._end &&
-        sourceUrl == other.sourceUrl;
-  }
+  bool operator ==(other) => other is _ScriptSpan
+      ? _startPosition == other._startPosition &&
+            _endPosition == other._endPosition &&
+            sourceUrl == other.sourceUrl
+      : super == other;
 
   FileSpan expand(FileSpan other) {
     if (sourceUrl != other.sourceUrl) {
@@ -304,12 +355,14 @@ class _ScriptSpan extends SourceSpanMixin implements FileSpan {
     }
 
     if (other is _ScriptSpan) {
-      var start = math.min(this._start, other._start);
-      var end = math.max(this._end, other._end);
-      return new _ScriptSpan(_script, start, end);
+      var startPosition = math.min(this._startPosition, other._startPosition);
+      var endPosition = math.max(this._endPosition, other._endPosition);
+      return new _ScriptSpan(_script, startPosition, endPosition,
+          startPosition == this._startPosition ? this._start : other._start,
+          endPosition == this._endPosition ? this._end : other._end);
     } else {
-      var start = math.min(this._start, other.start.offset);
-      var end = math.max(this._end, other.end.offset);
+      var start = math.min(this.start.offset, other.start.offset);
+      var end = math.max(this.end.offset, other.end.offset);
       return file.span(start, end);
     }
   }